Add files via upload
Finished creating a Twitter bot that searches for any tweets with the hashtag Eagle Creek Fire and or Breaking News and retweets it. I live in troutdale oregon just 2 minute walk west of 257th and am preparing for evacuation. This is to help me or others catch any breaking news we may miss